Zoom To implemented, validation + Minor Fixes
Added minor validation to coordinate system picker to prevent it
returning a blank when no rows are selected.  Added progress bar and
processing tracker label.   Added a extra detail to the log window.
Implemented "zoom to" checkbox so map zooms to the location of the new
layer (assuming new layer added to existing mapper)